# Conflicts:
#	src/main/java/com/ebicep/warlords/pve/upgrades/berserker/WoundingStrikeBranchBers.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/paladin/avenger/AvengerStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/paladin/avenger/AvengersWrathBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/paladin/crusader/CrusadersStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/paladin/protector/ProtectorStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/rogue/apothecary/ImpalingStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/rogue/assassin/JudgementStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/rogue/vindicator/RighteousStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/shaman/earthwarden/EarthenSpikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/shaman/thunderlord/LightningBoltBranch.java

+++ src/main/java/com/ebicep/warlords/abilties/GroundSlam.java
@@ -29,7 +29,7 @@ public class GroundSlam extends AbstractAbility {
 
     private int slamSize = 6;
 
-    public GroundSlam(String name, float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier) {
+    public GroundSlam(String name, float minDamageHeal, float maxDamageHeal, float cooldown, float energyCost, float critChance, float critMultiplier) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier);
     }
