+++ src/main/java/com/ebicep/warlords/abilties/SeismicWave.java
@@ -28,7 +28,7 @@ public class SeismicWave extends AbstractAbility {
     private int waveSize = 8;
     private int waveWidth = 1;
 
-    public SeismicWave(String name, float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier) {
+    public SeismicWave(String name, float minDamageHeal, float maxDamageHeal, float cooldown, float energyCost, float critChance, float critMultiplier) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier);
     }
