+++ src/main/java/com/ebicep/warlords/abilties/internal/AbstractStrikeBase.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 
 public abstract class AbstractStrikeBase extends AbstractAbility {
 
-    public AbstractStrikeBase(String name, float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier) {
+    public AbstractStrikeBase(String name, float minDamageHeal, float maxDamageHeal, float cooldown, float energyCost, float critChance, float critMultiplier) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier);
     }
